Refactored shared folder usage, using same dir structure in docker
