+++ src/oracle/bncWatcher/Dockerfile
@@ -9,6 +9,7 @@ COPY ./bncWatcher/package.json /watcher/
 
 RUN npm install
 
-COPY ./bncWatcher/bncWatcher.js ./shared/db.js ./shared/logger.js ./shared/crypto.js ./shared/amqp.js ./shared/wait.js ./shared/binanceClient.js /watcher/
+COPY ./bncWatcher/bncWatcher.js /watcher/src/
+COPY ./shared/db.js ./shared/logger.js ./shared/crypto.js ./shared/amqp.js ./shared/wait.js ./shared/binanceClient.js /watcher/shared/
 
-ENTRYPOINT ["node", "bncWatcher.js"]
+ENTRYPOINT ["node", "src/bncWatcher.js"]
