+++ src/oracle/bncWatcher/package.json
@@ -13,7 +13,10 @@
   },
   "engines": {
     "node": ">=10.6.0"
-  }
+  },
+  "files": [
+    "../shared"
+  ]
 }
