+++ src/oracle/ethWatcher/package.json
@@ -13,5 +13,8 @@
   },
   "engines": {
     "node": ">=10.6.0"
-  }
+  },
+  "files": [
+    "../shared"
+  ]
 }
