+++ src/oracle/proxy/Dockerfile
@@ -6,6 +6,7 @@ COPY ./proxy/package.json /proxy/
 
 RUN npm install
 
-COPY ./proxy/index.js ./proxy/encode.js ./proxy/decode.js ./proxy/sendTx.js ./proxy/contractsAbi.js ./proxy/utils.js ./proxy/expressUtils.js ./shared/logger.js ./shared/crypto.js ./shared/wait.js ./shared/binanceClient.js /proxy/
+COPY ./proxy/index.js ./proxy/encode.js ./proxy/decode.js ./proxy/sendTx.js ./proxy/contractsAbi.js ./proxy/utils.js ./proxy/expressUtils.js  /proxy/src/
+COPY ./shared/logger.js ./shared/crypto.js ./shared/wait.js ./shared/binanceClient.js /proxy/shared/
 
-ENTRYPOINT ["node", "index.js"]
+ENTRYPOINT ["node", "src/index.js"]
