+++ src/oracle/proxy/package.json
@@ -14,5 +14,8 @@
   },
   "engines": {
     "node": ">=10.6.0"
-  }
+  },
+  "files": [
+    "../shared"
+  ]
 }
