+++ src/oracle/tss-keygen/package.json
@@ -11,5 +11,8 @@
   },
   "engines": {
     "node": ">=10.6.0"
-  }
+  },
+  "files": [
+    "../shared"
+  ]
 }
