+++ src/oracle/tss-sign/package.json
@@ -12,5 +12,8 @@
   },
   "engines": {
     "node": ">=10.6.0"
-  }
+  },
+  "files": [
+    "../shared"
+  ]
 }
